Most Models/Textures Added
Added a lot of things to level 1. Created 4 different prefabs of trees and stones to be randomized.

Also animated the small slime sphere to go up and down.

Added water plane to the scene

Props on the left of the scene in folder "Single props" are the basics of each prefab and originals.

Floor and Walls still need to be textured.